+++ scripts/winesapos-upgrade.sh
@@ -916,6 +916,20 @@ ${CMD_PACMAN} -Q ceph-bin
 if [ $? -ne 0 ]; then
     ${CMD_YAY_INSTALL[*]} ceph-libs-bin ceph-bin
 fi
+
+grep -q "precedence ::ffff:0:0/96  100" /etc/gai.conf
+if [ $? -ne 0 ]; then
+    echo "label  ::1/128       0
+label  ::/0          1
+label  2002::/16     2
+label ::/96          3
+label ::ffff:0:0/96  4
+precedence  ::1/128       50
+precedence  ::/0          40
+precedence  2002::/16     30
+precedence ::/96          20
+precedence ::ffff:0:0/96  100" > /etc/gai.conf
+fi
 echo "Running 3.4.0 to 4.0.0 upgrades complete."
 
 echo "Upgrading system packages..."